disable kgdboc on ttyAMA0 when enable uart
The default `cmdline.txt` enables the kgdb on AMA0. 
When `enable_uart=1` appends in config.txt, the kernel hang up at boottime.

Since the user may have no monitor, the reason why the rpi hang up by enabling uart
is really hard to know. It causes me about 4 hours to find out the problem and solve it.

This patch adds the tips in the wiki, and hopes to save some time for other folks.